Extracts BottomBar & NavigationRail from ArmoryScreen
